V0.0.0.4
	修改imi-sh.c文件，增加保存配置的命令（增加客户端保存和IMISH保存命令）
		增加定时检测连接是否断开的操作（检测进程是否被重启，进程重启后IMISH需要重新连接）
	修改command.c文件，在保存配置的命令增加检测配置文件，备份配置文件是否存在的操作，
	优化ospf_dcn.c文件，删除LSA的时候确保把标识位清除，邻居状态发生变化的时候，确保没有安装的LSA
		进行安装
	修改watchquagga.c文件，修改设置log文件信息，
	修改make-setup文件，增加watchquagga模块功能，优化相关脚本，增加rt-service.sh和service.sh
		脚本，用于关闭启动路由协议，service.sh用于watchquagga监控客户端路由的时候进行进程控制的脚本

+++ lib/command.c
@@ -3095,21 +3095,21 @@ DEFUN (config_write_file,
       }
   vty_close (file_vty);
 
-  if (unlink (config_file_sav) != 0)
+  if (0 == access(config_file_sav,0) && unlink (config_file_sav) != 0)
     if (errno != ENOENT)
       {
 	vty_out (vty, "Can't unlink backup configuration file %s.%s", config_file_sav,
 		 VTY_NEWLINE);
         goto finished;
       }
-  if (link (config_file, config_file_sav) != 0)
+  if (0 == access(config_file_sav,0) && link (config_file, config_file_sav) != 0)
     {
       vty_out (vty, "Can't backup old configuration file %s.%s", config_file_sav,
 	        VTY_NEWLINE);
       goto finished;
     }
   sync ();
-  if (unlink (config_file) != 0)
+  if (0 == access(config_file_sav,0) && unlink (config_file) != 0)
     {
       vty_out (vty, "Can't unlink configuration file %s.%s", config_file,
 	        VTY_NEWLINE);
